+++ fotoskazka/app/Http/Controllers/CabinetController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Support\Facades\Auth;
+
+class CabinetController extends Controller
+{
+    public function index()
+    {
+        $user = Auth::user();
+
+        return view('cabinet.index', compact('user'));
+    }
+}