+++ fotoskazka/routes/auth.php
@@ -1,0 +1,10 @@
+<?php
+
+use App\Http\Controllers\Auth\LoginController;
+use Illuminate\Support\Facades\Route;
+
+Route::controller(LoginController::class)->group(function () {
+    Route::get('/login', 'create')->middleware('guest')->name('login');
+    Route::post('/login', 'store')->middleware('guest');
+    Route::post('/logout', 'destroy')->middleware('auth')->name('logout');
+});