+++ fotoskazka/routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\CabinetController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\ServiceController;
 use Illuminate\Support\Facades\Route;
@@ -18,3 +19,9 @@ Route::redirect('/blog', '/')->name('blog.index');
 Route::redirect('/blog/{slug}', '/')->name('blog.show');
 
 Route::post('/inquiry', [HomeController::class, 'storeInquiry'])->name('inquiry.store');
+
+Route::get('/cabinet', [CabinetController::class, 'index'])
+    ->middleware('auth')
+    ->name('cabinet.index');
+
+require __DIR__.'/auth.php';